requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
configparser==6.0.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.4
django-channels==0.7.0
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
gevent==23.9.0.post1
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
msgpack==1.0.5
numpy==1.25.2
oauthlib==3.2.2
pandas==2.0.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
service-identity==23.1.0
shutils==0.1.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
tzdata==2023.3
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
websocket==0.2.1
websockets==11.0.3
zope.event==5.0
zope.interface==6.0
